% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_rare_curve.R
\name{alpha_rare_curve}
\alias{alpha_rare_curve}
\title{Plotting rarefaction curve for each group}
\usage{
alpha_rare_curve(alpha_rare, metadata, groupID = "Group")
}
\arguments{
\item{alpha_rare}{alpha rarefaction matrix, typical output of usearch -alpha_div_rare;}

\item{metadata}{matrix or dataframe, including sampleID and groupID;}

\item{groupID}{column name for groupID.}
}
\value{
ggplot2 object.
}
\description{
Input alpha rare and metadata, and manual set metadata column names.
ggplot2 show line plot, and standard error
}
\details{
By default, returns grouped curve, sample parameter to be continued
}
\examples{
# usearch -alpha_div_rare result(alpha_rare), metadata and groupID
alpha_rare_curve(alpha_rare, metadata, "Group")
# alpha_rare, metadata, and groupID using site
alpha_rare_curve(alpha_rare, metadata, groupID="Site")
}
\references{
Yong-Xin Liu, Yuan Qin, Tong Chen, Meiping Lu, Xubo Qian, Xiaoxuan Guo & Yang Bai.
A practical guide to amplicon and metagenomic analysis of microbiome data.
Protein Cell, 2020(41), 1-16, DOI: \url{https://doi.org/10.1007/s13238-020-00724-8}

Jingying Zhang, Yong-Xin Liu, Na Zhang, Bin Hu, Tao Jin, Haoran Xu, Yuan Qin, Pengxu Yan, Xiaoning Zhang, Xiaoxuan Guo, Jing Hui, Shouyun Cao, Xin Wang, Chao Wang, Hui Wang, Baoyuan Qu, Guangyi Fan, Lixing Yuan, Ruben Garrido-Oter, Chengcai Chu & Yang Bai.
NRT1.1B is associated with root microbiota composition and nitrogen use in field-grown rice.
Nature Biotechnology, 2019(37), 6:676-684, DOI: \url{https://doi.org/10.1038/s41587-019-0104-4}
}
\seealso{
alpha_boxplot
}
\author{
Contact: Yong-Xin Liu \email{metagenome@126.com}
}
