% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_sample_rare.R
\name{alpha_sample_rare}
\alias{alpha_sample_rare}
\title{Plotting sample rarefaction boxplot}
\usage{
alpha_sample_rare(otutab, length = 18, rep = 30, count_cutoff = 1)
}
\arguments{
\item{otutab}{OTU table}

\item{length}{Number of boxplot, recommend < 30 & sample number}

\item{rep}{replicate sample times, is the dot number in boxplot, default 30, recommend < 30 & sample number}

\item{count_cutoff}{cutoff of reads count , default is 1}
}
\value{
ggplot2 object.
}
\description{
Input feature (OTU/ASV) table
}
\details{
By default, returns ggplot2
}
\examples{
# feature table, all parameters in default. Length is sample size times(boxplot number), rep is sampling times (dots in boxplot), count_cutoff is minimum reads count as detectable feature
alpha_sample_rare(otutab, length=18, rep=30, count_cutoff=1)
# Change the number of boxplot
alpha_sample_rare(otutab, length=9)
# Cutoff apparently affect trends
alpha_sample_rare(otutab, count_cutoff=9)
}
\references{
Yong-Xin Liu, Yuan Qin, Tong Chen, Meiping Lu, Xubo Qian, Xiaoxuan Guo & Yang Bai.
A practical guide to amplicon and metagenomic analysis of microbiome data.
Protein Cell, 2020(41), 1-16, DOI: \url{https://doi.org/10.1007/s13238-020-00724-8}

Jingying Zhang, Yong-Xin Liu, Na Zhang, Bin Hu, Tao Jin, Haoran Xu, Yuan Qin, Pengxu Yan, Xiaoning Zhang, Xiaoxuan Guo, Jing Hui, Shouyun Cao, Xin Wang, Chao Wang, Hui Wang, Baoyuan Qu, Guangyi Fan, Lixing Yuan, Ruben Garrido-Oter, Chengcai Chu & Yang Bai.
NRT1.1B is associated with root microbiota composition and nitrogen use in field-grown rice.
Nature Biotechnology, 2019(37), 6:676-684, DOI: \url{https://doi.org/10.1038/s41587-019-0104-4}
}
\seealso{
alpha_boxplot alpha_rare_curve alpha_rare_all
}
\author{
Contact: Yong-Xin Liu \email{metagenome@126.com}
}
