% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_cpcoa_dis.R
\name{beta_cpcoa_dis}
\alias{beta_cpcoa_dis}
\title{Plotting beta diversity scatter plot of Constrained PCoA}
\usage{
beta_cpcoa_dis(
  distance_mat,
  metadata,
  groupID = "genotype",
  ellipse = T,
  label = F
)
}
\arguments{
\item{metadata}{matrix or dataframe, including sampleID and groupID;}

\item{groupID}{column name for groupID.}

\item{ellipse}{stat ellipse, T or F.}

\item{label}{sample name showing, T or F.}

\item{dis}{distance type for caucluate, default "bray_curtis", alternative "unifrac, unifrac_binary, jaccard、manhatten, euclidean".}
}
\value{
ggplot2 object.
}
\description{
Input distance matrix and metadata, and manual set metadata column names.
ggplot2 show CPCoA with color and stat ellipse.
}
\details{
By default, returns beta CPCoA coordinate
The available diversity indices include the following:
\itemize{
\item{most used indices: bray_curtis, unifrac, unifrac_binary, jaccard}
\item{other used indices: manhatten, euclidean}
}
}
\examples{
# example data: OTU table, rownames is OTU_xxx, colnames is SampleID
data(beta_bray_curtis)
# example data: metadata or design, include SampleID, genotype and site
data(metadata)
# Set 2 parameters: otu table, metadata, and distance type and groupID using default "bray_curtis" and "genotype"
beta_cpcoa_dis(beta_bray_curtis, metadata)
# Set 3 parameters: distance matrix, metadata, and groupID as "site"
beta_cpcoa_dis(beta_bray_curtis, metadata, "site")
# Set 5 parameters: distance matrix, metadata, and groupID as using "genotype", open stat elipse and sample label
beta_cpcoa_dis(beta_bray_curtis, metadata, groupID = "genotype", ellipse = T, label = T)
}
\references{
Yong-Xin Liu, Yuan Qin, Tong Chen, Meiping Lu, Xubo Qian, Xiaoxuan Guo & Yang Bai.
A practical guide to amplicon and metagenomic analysis of microbiome data.
Protein Cell, 2020, DOI: \url{https://doi.org/10.1007/s13238-020-00724-8}

Jingying Zhang, Yong-Xin Liu, Na Zhang, Bin Hu, Tao Jin, Haoran Xu, Yuan Qin, Pengxu Yan, Xiaoning Zhang, Xiaoxuan Guo, Jing Hui, Shouyun Cao, Xin Wang, Chao Wang, Hui Wang, Baoyuan Qu, Guangyi Fan, Lixing Yuan, Ruben Garrido-Oter, Chengcai Chu & Yang Bai.
NRT1.1B is associated with root microbiota composition and nitrogen use in field-grown rice.
Nature Biotechnology, 2019(37), 6:676-684, DOI: \url{https://doi.org/10.1038/s41587-019-0104-4}
}
\seealso{
beta_cpcoa_dis
}
\author{
Contact: Yong-Xin Liu \email{metagenome@126.com}
}
