% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_pcoa.R
\name{beta_pcoa}
\alias{beta_pcoa}
\title{Plotting beta diversity scatter plot}
\usage{
beta_pcoa(
  dis_mat,
  metadata,
  groupID = "Group",
  ellipse = T,
  label = F,
  PCo = 12
)
}
\arguments{
\item{dis_mat}{distance matrix, typical output of usearch -beta_div,}

\item{metadata}{matrix or dataframe, including sampleID and groupID;}

\item{groupID}{column name for groupID.}

\item{ellipse}{stat ellipse, T or F.}

\item{label}{sample name showing, T or F.}

\item{PCo}{principle coordinate used, default 12, alternative 13, or 23.}
}
\value{
ggplot2 object.
}
\description{
Input distance matrix and metadata, and manual set metadata column names.
Visualize PCoA with color and stat ellipse by ggplot2.
}
\details{
By default, returns beta PCoA coordinate
The available diversity indices include the following:
\itemize{
\item{most used indices: bray_curtis, unifrac}
\item{other used indices: unifrac_binary, jaccard, euclidean, manhatten}
}
}
\examples{
# Set essential 3 parameters: distance matrix, metadata and groupID
beta_pcoa(beta_bray_curtis, metadata, "Group")
# Set full 6 parameters: distance matrix, metadata, and groupID as using "site",
beta_pcoa(dis_mat=beta_unifrac, metadata=metadata, groupID="Site", ellipse=F, label=T, PCo=13)
}
\references{
Yong-Xin Liu, Yuan Qin, Tong Chen, Meiping Lu, Xubo Qian, Xiaoxuan Guo & Yang Bai.
A practical guide to amplicon and metagenomic analysis of microbiome data.
Protein Cell, 2020(41), 1-16, DOI: \url{https://doi.org/10.1007/s13238-020-00724-8}

Jingying Zhang, Yong-Xin Liu, Na Zhang, Bin Hu, Tao Jin, Haoran Xu, Yuan Qin, Pengxu Yan, Xiaoning Zhang, Xiaoxuan Guo, Jing Hui, Shouyun Cao, Xin Wang, Chao Wang, Hui Wang, Baoyuan Qu, Guangyi Fan, Lixing Yuan, Ruben Garrido-Oter, Chengcai Chu & Yang Bai.
NRT1.1B is associated with root microbiota composition and nitrogen use in field-grown rice.
Nature Biotechnology, 2019(37), 6:676-684, DOI: \url{https://doi.org/10.1038/s41587-019-0104-4}
}
\seealso{
beta_cpcoa
}
\author{
Contact: Yong-Xin Liu \email{metagenome@126.com}
}
