% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetaDiv.R
\name{BetaDiv}
\alias{BetaDiv}
\title{Beta diversity plotting}
\usage{
BetaDiv(
  otu = otutab,
  map = metadata,
  tree = tree,
  ps = NULL,
  group = "Group",
  dist = "bray",
  method = "PCoA",
  Micromet = "adonis",
  pvalue.cutoff = 0.05
)
}
\arguments{
\item{otu}{OTU/ASV table;}

\item{map}{Sample metadata;}

\item{tree}{tree/nwk file;}

\item{group}{group ID;}

\item{dist}{distance type, including "unifrac" "wunifrac" "dpcoa" "jsd" "manhattan" "euclidean"   "canberra" "bray" "kulczynski"  "jaccard" "gower" "altGower" "morisita" "horn" "mountford"  "raup" "binomial"  "chao"  "cao" "w"  "-1"  "c" "wb"  "r"   "I"  "e" "t" "me"   "j"  "sor"  "m"   "-2"  "co";}

\item{method}{DCA, CCA, RDA, NMDS, MDS, PCoA, PCA, LDA;}

\item{Micromet}{statistics by adonis/anosim/MRPP;}

\item{pvalue.cutoff}{Pvalue threshold;}
}
\value{
list object including plot, stat table
}
\description{
Input otutab, metadata and tree or phyloseq object; support 47 distance type (bray, unifrac, wunifrac ...),  8 ordination method (PCoA, NMDS, ...); output ggplot2 figure, data and statistical test result.
}
\details{
By default, input phyloseq object include metadata, otutab and tree
The available diversity indices include the following:
\itemize{
\item{most used indices: bray, unifrac, wunifrac}
\item{other used indices: manhattan, euclidean, jaccard ...}
}
}
\examples{

BetaDiv(otu = otutab_rare, map = metadata, tree = tree, group = "Group", dist = "bray", method = "PCoA", Micromet = "adonis", pvalue.cutoff = 0.05)

}
\references{
Jingying Zhang, Yong-Xin Liu, Na Zhang, Bin Hu, Tao Jin, Haoran Xu, Yuan Qin, Pengxu Yan, Xiaoning Zhang, Xiaoxuan Guo, Jing Hui, Shouyun Cao, Xin Wang, Chao Wang, Hui Wang, Baoyuan Qu, Guangyi Fan, Lixing Yuan, Ruben Garrido-Oter, Chengcai Chu & Yang Bai.
NRT1.1B is associated with root microbiota composition and nitrogen use in field-grown rice.
Nature Biotechnology, 2019(37), 6:676-684, DOI: \url{https://doi.org/10.1038/s41587-019-0104-4}
}
\seealso{
beta_pcoa beta_cpcoa
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn}, Yong-Xin Liu \email{yxliu@genetics.ac.cn}
}
