% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format2lefse.R
\name{format2lefse}
\alias{format2lefse}
\title{Format otutab and taxonomy into each level for STAMP}
\usage{
format2lefse(otutab, taxonomy, metadata, thre = 0.01,
  groupID = "Group", output = "LEfSe.txt")
}
\arguments{
\item{otutab}{row or normalized OTU table}

\item{taxonomy}{Taxonomy include seven taxonomy level in tsv format}

\item{metadata}{matrix or dataframe, including sampleID and groupID}

\item{thre}{threshold of OTU abundance, especially for lefse create properly cladegram}

\item{groupID}{column name for groupID.}
}
\value{
nothing, all table 0-8 saving in same directory.
}
\description{
Input otutab, taxonomy and metadata, and manual set abundance threshold, metadata column names.
dplyr merge taxonomy.
}
\details{
By default, written 1 file
}
\examples{
# Set six parameters: three input files, and three adjust parameters
format2lefse(otutab, taxonomy, metadata, thre = 0.01, groupID = "Group")
}
\references{
Zhang, J., Zhang, N., Liu, YX. et al.
Root microbiota shift in rice correlates with resident time in the field and developmental stage
Sci. China Life Sci. (2018) 61: 613.  DOI: \url{https://doi.org/10.1007/s11427-018-9284-4}
}
\seealso{
alpha_rare
}
\author{
Contact: Yong-Xin Liu \email{metagenome@126.com}
}
