% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format2stamp.R
\name{format2stamp}
\alias{format2stamp}
\title{Format otutab and taxonomy into each level for STAMP}
\usage{
format2stamp(otutab, taxonomy, thre = 1e-04, prefix = "tax")
}
\arguments{
\item{otutab}{row or normalized OTU table}

\item{taxonomy}{Taxonomy include seven taxonomy level in tsv format}

\item{thre}{threshold of OTU abundance, default 1 per Million}

\item{prefix}{prefix of output files}
}
\value{
nothing, all table 1-8 saving in same directory.
}
\description{
Input otutab, taxonomy, and manual set abundance threshold.
dplyr merge taxonomy.
}
\details{
By default, written 9 files
The files list as the following:
\itemize{
\item{1-7: Kingdom to Species 7 levels}
\item{8: filtered OTUs}
}
}
\examples{
# Set 4 parameters: 2 input files, and 2 adjust parameters
format2stamp(otutab, taxonomy, thre = 0.01,  prefix = "tax")
}
\references{
Zhang, J., Zhang, N., Liu, YX. et al.
Root microbiota shift in rice correlates with resident time in the field and developmental stage
Sci. China Life Sci. (2018) 61: 613.  DOI: \url{https://doi.org/10.1007/s11427-018-9284-4}
}
\seealso{
alpha_rare
}
\author{
Contact: Yong-Xin Liu \email{metagenome@126.com}
}
