% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_stackplot.R
\name{func_stackplot}
\alias{func_stackplot}
\title{Plotting stackplot of function for groups or samples}
\usage{
func_stackplot(
  func_sum,
  metadata,
  topN = 8,
  groupID = "Group",
  style = "group",
  sorted = "abundance"
)
}
\arguments{
\item{func_sum}{composition matrix, like OTU table and rowname is function, typical output of usearch -sinfunc_summary;}

\item{metadata}{matrix or dataframe, including sampleID and groupID;}

\item{topN}{Top N function to show, default 8, alternative 4, 6, 10 ...;}

\item{groupID}{column name for groupID;}

\item{style}{group or sample, default group}

\item{sorted}{Legend sorted type, default abundance, alternative alphabet}
}
\value{
ggplot2 object.
}
\description{
Input function composition, and metadata (SampleID and groupID). Then select top N high abundance function and group other low abundance. When Select samples can draw sample composition by facet groups. If used group can show mean of each group. Finally, return a ggplot2 object.
}
\details{
By default, returns top 8 function and group mean stackplot
The available style include the following:
\itemize{
\item{group: group mean stackplot}
\item{sample: each sample stackplot and facet by group}
}
}
\examples{
# Function table in phylum level, rownames is Phylum, colnames is SampleID
data(func_phylum)
# metadata, include SampleID, Group and Site
data(metadata)
# func_sum and metadata as input, default include top 8 function, groupID is Group, show group mean and sorted by abundance
func_stackplot(func_sum = func_phylum, metadata)
# Set top10 function, group by "Site", group mean, and sort by abundance
func_stackplot(func_sum = func_phylum, metadata, topN = 10, groupID = "Site", style = "group", sorted = "abundance")
# Set top 10 function, group by "Group", and sample composition sorted by alphabet
func_stackplot(func_sum = func_phylum, metadata, topN = 10, groupID = "Group", style = "sample", sorted = "alphabet")
}
\references{
Yong-Xin Liu, Yuan Qin, Tong Chen, Meiping Lu, Xubo Qian, Xiaoxuan Guo & Yang Bai.
A practical guide to amplicon and metagenomic analysis of microbiome data.
Protein Cell, 2020, DOI: \url{https://doi.org/10.1007/s13238-020-00724-8}

Jingying Zhang, Yong-Xin Liu, Na Zhang, Bin Hu, Tao Jin, Haoran Xu, Yuan Qin, Pengxu Yan, Xiaoning Zhang, Xiaoxuan Guo, Jing Hui, Shouyun Cao, Xin Wang, Chao Wang, Hui Wang, Baoyuan Qu, Guangyi Fan, Lixing Yuan, Ruben Garrido-Oter, Chengcai Chu & Yang Bai.
NRT1.1B is associated with root microbiota composition and nitrogen use in field-grown rice.
Nature Biotechnology, 2019(37), 6:676-684, DOI: \url{https://doi.org/10.1038/s41587-019-0104-4}
}
\seealso{
func_stackplot
}
\author{
Contact: Yong-Xin Liu \email{metagenome@126.com}
}
