% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_rda-cca.R
\name{RDA_CCA}
\alias{RDA_CCA}
\title{RDA or CCA calculate}
\usage{
RDA_CCA(otu = otutab, map = metadata, ps = NULL, env = env, group = "Group")
}
\arguments{
\item{otu}{OTU/ASV table;}

\item{map}{Sample metadata;}

\item{env}{environmental indicators table;}

\item{group}{group ID;}
}
\value{
list object including plot, stat table
}
\description{
Input otutab, metadata and environmental factors or phyloseq object; support Automatic select which method : RDA or CCA were used for analyze; output ggplot2 plot, data and statistical test result.
}
\details{
By default, input phyloseq object include metadata, otutab and env table
}
\examples{

RDA_CCA(otu=otutab, map=metadata, ps=NULL, env=env, group="Group")

}
\references{
Yong-Xin Liu, Yuan Qin, Tong Chen, Meiping Lu, Xubo Qian, Xiaoxuan Guo & Yang Bai.
A practical guide to amplicon and metagenomic analysis of microbiome data.
Protein Cell, 2020(41), 1-16, DOI: \url{https://doi.org/10.1007/s13238-020-00724-8}

Jingying Zhang, Yong-Xin Liu, Na Zhang, Bin Hu, Tao Jin, Haoran Xu, Yuan Qin, Pengxu Yan, Xiaoning Zhang, Xiaoxuan Guo, Jing Hui, Shouyun Cao, Xin Wang, Chao Wang, Hui Wang, Baoyuan Qu, Guangyi Fan, Lixing Yuan, Ruben Garrido-Oter, Chengcai Chu & Yang Bai.
NRT1.1B is associated with root microbiota composition and nitrogen use in field-grown rice.
Nature Biotechnology, 2019(37), 6:676-684, DOI: \url{https://doi.org/10.1038/s41587-019-0104-4}
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn}, Yong-Xin Liu \email{yxliu@genetics.ac.cn}
}
