% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_maptree.R
\name{tax_maptree}
\alias{tax_maptree}
\title{Plotting maptree of taxonomy}
\usage{
tax_maptree(x)
}
\arguments{
\item{mapdata}{format2maptree result}
}
\value{
ggplot2 object.
}
\description{
Visualize mapdata. Finally, return a ggplot2 object.
}
\details{
By default, returns  mapadd[\link{1}] is ggplot object
The available style include the following:
\itemize{
\item{phylum: color by phylum}
\item{diff: color by diff}
}
}
\examples{
# Input feature table, taxonomy and Top N features, and format into mapdata
mapdata = format2maptree(otutab, taxonomy, 200)
# Add mean abundance size and phylum color for maptree
mapadd = tax_maptree(mapdata)
# Saving and plotting maptree
(p = mapadd[[1]])
}
\references{
Yong-Xin Liu, Yuan Qin, Tong Chen, Meiping Lu, Xubo Qian, Xiaoxuan Guo & Yang Bai.
A practical guide to amplicon and metagenomic analysis of microbiome data.
Protein Cell, 2020, DOI: \url{https://doi.org/10.1007/s13238-020-00724-8}

Jingying Zhang, Yong-Xin Liu, Na Zhang, Bin Hu, Tao Jin, Haoran Xu, Yuan Qin, Pengxu Yan, Xiaoning Zhang, Xiaoxuan Guo, Jing Hui, Shouyun Cao, Xin Wang, Chao Wang, Hui Wang, Baoyuan Qu, Guangyi Fan, Lixing Yuan, Ruben Garrido-Oter, Chengcai Chu & Yang Bai.
NRT1.1B is associated with root microbiota composition and nitrogen use in field-grown rice.
Nature Biotechnology, 2019(37), 6:676-684, DOI: \url{https://doi.org/10.1038/s41587-019-0104-4}
}
\seealso{
data_to_maptree
}
\author{
Contact: Tao Wen, Yong-Xin Liu \email{metagenome@126.com}
}
